feat: сортирует комментарии и обрезает их отображение до 10
