+++ src/hooks/use-fetch-one-offer.ts
@@ -2,7 +2,6 @@ import {useEffect} from 'react';
 import {oneOfferActions} from '../redux/slices/one-offer-slice';
 import {RequestStatus} from '../types/redux-types';
 import {useActionCreators} from './redux-hooks';
-import {OfferCard} from '../types/common-types';
 
 export function useFetchOneOffer(status: RequestStatus, id: string) {
   const {fetchOneOffer, fetchNearestOffers, clear} = useActionCreators(oneOfferActions);
@@ -16,12 +15,4 @@ export function useFetchOneOffer(status: RequestStatus, id: string) {
   useEffect(() => () => {
     clear();
   }, [id, clear]);
-}
-
-export function sliceArrayToThreeElements(array: OfferCard[]) {
-  if (array.length > 3) {
-    const randomIndex = Math.floor(Math.random() * array.length);
-    return array.slice(randomIndex, randomIndex + 3);
-  }
-  return array;
 }
