--- src/mock/reviews.ts
@@ -1,24 +1,0 @@
-export const REVIEWS = [
-  {
-    'id': 'b67ddfd5-b953-4a30-8c8d-bd083cd6b62a',
-    'date': '2019-05-08T14:13:56.569Z',
-    'user': {
-      'name': 'Oliver Conner',
-      'avatarUrl': 'https://url-to-image/image.png',
-      'isPro': false
-    },
-    'comment': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
-    'rating': 4
-  },
-  {
-    'id': 'b67ddfd5-b953-4a30-8c8d-bd083cd6b62b',
-    'date': '2019-05-08T14:13:56.569Z',
-    'user': {
-      'name': 'Max',
-      'avatarUrl': 'https://url-to-image/image.png',
-      'isPro': false
-    },
-    'comment': 'A quiet place for a vacation.',
-    'rating': 3.9
-  }
-];
